refactor(switch): modularize and improve component structure
Decoupled logic, styles, and interface definitions for the Switch component into separate files. This improves maintainability, ensures better modularity, and aligns the codebase with a more scalable structure. Also updated references and renamed files to follow consistent naming conventions.
