--- src/components/selection/slider/index.tsx
@@ -1,2 +1,0 @@
-export * from './slider.component';
-export * from './SliderStyle';
