+++ src/components/selection/switch/index.ts
@@ -1,2 +1,3 @@
-export * from './Switch';
-export * from './SwitchStyle';
+export * from './switch.component';
+export * from './switch.interface';
+export * from './switch.style';
